+++ starters/solid/client/index.html
@@ -5,10 +5,10 @@
     <link rel="stylesheet" href="./base.css">
     <!-- head -->
     <!-- hydrationScript -->
-    <!-- hydration -->
   </head>
   <body>
     <main><!-- element --></main>
   </body>
+  <!-- hydration -->
   <script type="module" src="/dx:mount.js"></script>
 </html>
